Modify Range and Merge queries to track backend team OKRs
This commit modifies both the Range and Merge queries to do 15m and 7 day
queries (even though 15 min range is not tracked). This commit also changes
merge queries to run against a profile type instead of a specific series (which
is a more usual query to run).

Additionally a couple of cleanups were done:
- Removing executeLoop label to avoid goto statements.
- Logging was added/enhanced.