*: refactor
This commit refactors all querying logic into a new Querier struct. Since all
queries are more or less dependent on each other, the execution logic is
simplified to run the queries one at a time, with the option to add concurrent
Queriers in the future if we want to increase load. This also exposes a single
"query-interval" flag to make it easier for the user to choose a single
interval in which to execute queries.

Other bits and bobs were also improved:
- Termination has been simplified
- Expiring maps have been removed. Although it was a good approach to store
state needed by future queries, it is simpler and more memory efficient to
just store the last result for each query. Since queries are now run in the
same goroutine one after another, there is also no need for synchronization.
- Queries are no longer run conditionally. For example, we would avoid running
15 minute merges if QueryRange didn't return a series with enough data.
Anecdotally, this seemed broken and it was implemented in a leaky way since
we would always run a QueryRange and store the latency as valid even if the
data QueryRange executed on was less than the expected range. This simplifies
querying logic.